Fix numpy float64 type errors and yfinance 1.x compatibility
- Cast all values to Python float before DB insert; yfinance 1.x
  returns np.float64 which psycopg2 misinterprets as a schema name
- Add retry with backoff on yf.Ticker().info (handles 429 rate limits)
- Upgrade yfinance to 1.3.0 for better session/cookie handling
- load_dotenv now uses __file__-relative path so CLI works from any dir
- init_db auto-creates the local postgres database if missing

+++ analysis/metrics.py
@@ -6,7 +6,7 @@ from db.models import Stock, Financials, ComputedMetrics
 
 def _pct_change(new, old):
     if old and old != 0 and new is not None:
-        return round((new - old) / abs(old) * 100, 2)
+        return round(float((new - old) / abs(old) * 100), 2)
     return None
